add varying styles for feature-image component
