+++ apps/web/vibes/soul/docs/featured-image.mdx
@@ -1,5 +1,9 @@
 ---
 title: Featured Image
-preview: featured-image-example
+preview:
+  Electric: featured-image-electric-example
+  Warm: featured-image-warm-example
+  Blue: featured-image-blue-example
+  Luxury: featured-image-luxury-example
 previewSize: lg
 ---
