+++ apps/web/vibes/soul/examples.ts
@@ -123,11 +123,32 @@ export const examples = [
     component: lazy(() => import('./examples/featured-blog-post-list')),
   },
   {
-    name: 'featured-image-example',
+    name: 'featured-image-electric-example',
     dependencies: [],
     registryDependencies: [],
-    files: ['examples/featured-image.tsx'],
-    component: lazy(() => import('./examples/featured-image')),
+    files: ['examples/featured-image-electric.tsx'],
+    component: lazy(() => import('./examples/featured-image-electric')),
+  },
+  {
+    name: 'featured-image-warm-example',
+    dependencies: [],
+    registryDependencies: [],
+    files: ['examples/featured-image-warm.tsx'],
+    component: lazy(() => import('./examples/featured-image-warm')),
+  },
+  {
+    name: 'featured-image-blue-example',
+    dependencies: [],
+    registryDependencies: [],
+    files: ['examples/featured-image-blue.tsx'],
+    component: lazy(() => import('./examples/featured-image-blue')),
+  },
+  {
+    name: 'featured-image-luxury-example',
+    dependencies: [],
+    registryDependencies: [],
+    files: ['examples/featured-image-luxury.tsx'],
+    component: lazy(() => import('./examples/featured-image-luxury')),
   },
   {
     name: 'featured-products-carousel-example',
